backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart
email-validator
aiosmtplib
jinja2